refactoring google maps service out
