+++ src/app/home/home.page.ts
@@ -1,10 +1,9 @@
 import { Component, ElementRef } from '@angular/core';
 
-import { Geolocation } from '@ionic-native/geolocation/ngx';
-import { NativeGeocoder, NativeGeocoderResult, NativeGeocoderOptions } from '@ionic-native/native-geocoder/ngx';
 import { ViewChild } from '@angular/core'
+import { GoogleApiService } from '../google-api.service';
 
-declare var google;
+
 
 @Component({
   selector: 'app-home',
@@ -16,66 +15,19 @@ export class HomePage {
   @ViewChild('map',{static:false}) mapElement: ElementRef;
   map: any;
   address:string;
+  google_api: GoogleApiService;
 
   constructor(
-    private geolocation: Geolocation,
-    private nativeGeocoder: NativeGeocoder) {
+    google_api: GoogleApiService
+  ) {
+    this.google_api = google_api;
   }
  
  
-  ngOnInit() {
-    this.loadMap();
+  ngAfterViewInit() {
+    this.google_api.loadMap(this.mapElement);
   }
  
-  loadMap() {
-    this.geolocation.getCurrentPosition().then((resp) => {
-      let latLng = new google.maps.LatLng(resp.coords.latitude, resp.coords.longitude);
-      let mapOptions = {
-        center: latLng,
-        zoom: 15,
-        mapTypeId: google.maps.MapTypeId.ROADMAP
-      }
- 
-      this.getAddressFromCoords(resp.coords.latitude, resp.coords.longitude);
- 
-      this.map = new google.maps.Map(this.mapElement.nativeElement, mapOptions);
- 
-      this.map.addListener('tilesloaded', () => {
-        console.log('accuracy',this.map);
-        this.getAddressFromCoords(this.map.center.lat(), this.map.center.lng())
-      });
- 
-    }).catch((error) => {
-      console.log('Error getting location', error);
-    });
-  }
- 
-  getAddressFromCoords(lattitude, longitude) {
-    console.log("getAddressFromCoords "+lattitude+" "+longitude);
-    let options: NativeGeocoderOptions = {
-      useLocale: true,
-      maxResults: 5
-    };
- 
-    this.nativeGeocoder.reverseGeocode(lattitude, longitude, options)
-      .then((result: NativeGeocoderResult[]) => {
-        this.address = "";
-        let responseAddress = [];
-        for (let [key, value] of Object.entries(result[0])) {
-          if(value.length>0)
-          responseAddress.push(value);
- 
-        }
-        responseAddress.reverse();
-        for (let value of responseAddress) {
-          this.address += value+", ";
-        }
-        this.address = this.address.slice(0, -2);
-      })
-      .catch((error: any) =>{ 
-        this.address = "Address Not Available!";
-      });
- 
-  }
+  
 
 }
